Locker Scene, Title Screen, Player UI, added gamejam sounds 2 into Project
First Pass on everything, Locker is able to be opened and closed (right now it's just swaping cameras), it also has sounds. Title Screen has buttons and pannels laid out but only temp art/text in it. Player UI has a spot for notification text, and a first pass on the watch meter