Added whispering_area; Updated ParentArgumentv1
whispering area is a whispering ambience that should be kept lower in volume in relation to the overall mix (there's some shakespeare quoting in there...anyways I have some individual recordings too that I could add as separate files). The argument has kept the original name, which I hope would help,  but the volume is boosted.